add jwt authentication and authorization

+++ controllers/api/v1/posts_api.js
@@ -22,11 +22,18 @@ module.exports.index = async (req, res) => {
 module.exports.destroy = async (req, res) => {
     try {
         let post = await Post.findById(req.params.id);
-        post.remove();
-        await Comment.deleteMany({post: req.params.id});
-        return res.status(200).json({
-            message: "Post and associated comments deleted successfully"
-        });
+        if(post.user._id == req.user.id) {
+            post.remove();
+            await Comment.deleteMany({post: req.params.id});
+            return res.status(200).json({
+                message: "Post and associated comments deleted successfully"
+            });
+        } else {
+            return res.status(401).json({
+                message: "You cannot delete this post"
+            });
+        }
+        
     } catch(err) {
         console.log(err);
         return res.status(500).json({
